requirements.txt
pandas
numpy
scikit-learn
python-dotenv
langchain
langchain-community
pypdf
openai
chromadb
tiktoken
langchain-community
fastapi
langchain-openai
uvicorn
